Base

+++ src/camera.h
@@ -11,11 +11,11 @@ class Camera
 {
 public:
     explicit Camera(const float fov, const glm::vec3& pos, const glm::vec3& lookAt,
-                    const glm::vec3& up, const int width, const int height);
+                    const glm::vec3& up, const unsigned width, const unsigned height);
     
     void generateRay(const Sample& s, Ray* ray) const;
-    unsigned short width() const { return mWidth; }
-    unsigned short height() const { return mHeight; }
+    unsigned width() const { return mWidth; }
+    unsigned height() const { return mHeight; }
 
 private:
     explicit Camera() :
@@ -24,7 +24,7 @@ private:
     { } // Don't use this
 
     const glm::vec3 mPos, mLookAt, mUp;
-    const unsigned short mWidth, mHeight;
+    const unsigned mWidth, mHeight;
     const float mFov;
     float mAlpha, mBeta;
     glm::vec3 mU, mV, mW;
